Convert DICOM Header to Analyze

A subset of header information from DICOM is placed into Analyze 7.5 format.

dicom2analyze(dcm, datatype = 4, reslice = TRUE, DIM = 3,
  descrip = "SeriesDescription", ...)

Arguments

dcm
DICOM object containing both header and image information.

datatype
is an integer that denotes the type of data contained in each voxel. See convert.datatype.anlz or the ANALYZE documentation for more details.

reslice
Logical variable (default = TRUE) indicating if the data volume should be resliced.

DIM
The dimension of the array to be used (default = 3D).

descrip
DICOM header field(s) to be included in the descrip

...
Arguments to be passed to anlz

Value

An object of class anlz.

Details

See the references.

References

Analyze 7.5
https://rportal.mayo.edu/bir/ANALYZE75.pdf

Digital Imaging and Communications in Medicine (DICOM)
http://medical.nema.org

See also

Author

Brandon Whitcher bwhitcher@gmail.com

Examples

if (FALSE) {
dcmList <- dicomSeparate(system.file("hk-40", package="oro.dicom"))
require("oro.nifti")
dcmAnlz <- dicom2analyze(dcmList, datatype=4, mode="integer")
image(dcmAnlz)
orthographic(dcmAnlz)
}