Match String to DICOM Header Field

A convenient wrapper function that utilizes internal functions to match character strings with the DICOM header information.

matchHeader(hdr, string)

Arguments

hdr
is the result from extracting information from a DICOM header field; e.g., using extractHeader.

string
is a character string to be matched with the DICOM header.

Value

A logical vector of length length(hdr).

References

Digital Imaging and Communications in Medicine (DICOM)
http://medical.nema.org

See also

Author

Brandon Whitcher bwhitcher@gmail.com

Examples

x <- readDICOMFile(system.file("dcm/Abdo.dcm", package="oro.dicom"))
modality <- extractHeader(x$hdr, "Modality", numeric=FALSE)
matchHeader(modality, "mr") # case insensitive by default
#> [1] TRUE